Ensure builtin datapack loads at top priority
