+++ src/main/resources/the_expanse/data/the_expanse/dimension_type/overworld.json
@@ -1,0 +1,22 @@
+{
+  "ultrawarm": false,
+  "natural": true,
+  "coordinate_scale": 1.0,
+  "piglin_safe": false,
+  "bed_works": true,
+  "respawn_anchor_works": false,
+  "has_raids": true,
+  "min_y": -256,
+  "height": 2272,
+  "logical_height": 2272,
+  "infiniburn": "#minecraft:infiniburn_overworld",
+  "effects": "minecraft:overworld",
+  "ambient_light": 0.0,
+  "monster_spawn_light_level": {
+    "type": "minecraft:uniform",
+    "value": 0
+  },
+  "monster_spawn_block_light_limit": 0,
+  "has_skylight": true,
+  "sea_level": 150
+}
