+++ src/main/resources/the_expanse/data/the_expanse/dimension_type/the_nether.json
