+++ src/main/resources/the_expanse/data/the_expanse/worldgen/noise_settings/end_expanded.json
@@ -1,0 +1,20 @@
+{
+  "sea_level": 0,
+  "disable_mob_generation": false,
+  "aquifers_enabled": false,
+  "ore_veins_enabled": false,
+  "default_block": {
+    "Name": "minecraft:end_stone"
+  },
+  "default_fluid": {
+    "Name": "minecraft:air"
+  },
+  "noise": {
+    "min_y": -256,
+    "height": 2272,
+    "size_horizontal": 1,
+    "size_vertical": 2
+  },
+  "spawn_target": [],
+  "spawn_density": {}
+}
