+++ src/main/resources/the_expanse/data/the_expanse/worldgen/noise_settings/overworld_expanded.json
@@ -1,0 +1,20 @@
+{
+  "sea_level": 150,
+  "disable_mob_generation": false,
+  "aquifers_enabled": true,
+  "ore_veins_enabled": true,
+  "default_block": {
+    "Name": "minecraft:stone"
+  },
+  "default_fluid": {
+    "Name": "minecraft:water"
+  },
+  "noise": {
+    "min_y": -256,
+    "height": 2272,
+    "size_horizontal": 1,
+    "size_vertical": 2
+  },
+  "spawn_target": [],
+  "spawn_density": {}
+}
