+++ src/main/resources/the_expanse/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+  "pack": {
+    "pack_format": 48,
+    "description": "The Expanse builtin datapack: expanded dimensions, noise, and carvers"
+  }
+}
